Updated Constants location
