+++ components/onboarding/onboarding-concept-guides.tsx
@@ -1,40 +1,12 @@
 import { ExternalLink } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import { ONBOARDING_CONCEPT_GUIDES, OnboardingGuideKey } from "@/constants/onboarding.constants";
+import { VS_CODE_DOWNLOAD } from "@/constants/onboarding.constants";
 
 /**
  * Central place for “New to this?” / guide links. Import these in step content.
  * Paths resolve under /onboarding/guides (re-exporting onboarding-old content).
  */
-export const ONBOARDING_CONCEPT_GUIDES = {
-  github: {
-    href: "/onboarding/guides/github",
-    label: "Understanding GitHub",
-  },
-  git: {
-    href: "/onboarding/guides/git",
-    label: "Understanding Git",
-  },
-  vscode: {
-    href: "/onboarding/guides/vscode",
-    label: "Understanding VS Code",
-  },
-  discord: {
-    href: "/onboarding/guides/discord",
-    label: "Understanding Discord",
-  },
-  linkedin: {
-    href: "/onboarding/guides/linkedin",
-    label: "Understanding LinkedIn",
-  },
-  leetcode: {
-    href: "/onboarding/guides/leetcode",
-    label: "Understanding LeetCode",
-  },
-} as const;
-
-export type OnboardingGuideKey = keyof typeof ONBOARDING_CONCEPT_GUIDES;
-
-const VS_CODE_DOWNLOAD = "https://code.visualstudio.com/download";
 
 export function VsCodeDownloadLink() {
   return (
